# ARROW-6038: [C++] Faster type equality