--- a/cpp/src/arrow/CMakeLists.txt
+++ b/cpp/src/arrow/CMakeLists.txt
@@ -378,6 +378,7 @@ add_arrow_test(tensor-test)
 add_arrow_test(sparse_tensor-test)
 
 add_arrow_benchmark(builder-benchmark)
+add_arrow_benchmark(type-benchmark)
 
 add_subdirectory(array)
 add_subdirectory(csv)