… cpp/src/arrow/type-test.cc
… cpp/src/arrow/type.h