When checking for type equality, compute and cache a fingerprint of the type so as to avoid costly nested type walking and multiple comparisons.

Before:
```
----------------------------------------------------------------
Benchmark                         Time           CPU Iterations
----------------------------------------------------------------
TypeEqualsSimple                 13 ns         13 ns   55242976   150.558M items/s
TypeEqualsComplex               430 ns        430 ns    1637275   4.43634M items/s
TypeEqualsWithMetadata          595 ns        595 ns    1199216   3.20778M items/s
SchemaEquals                   1465 ns       1465 ns     479512   1.30226M items/s
SchemaEqualsWithMetadata        922 ns        922 ns     763752    2.0683M items/s
```

After:
```
----------------------------------------------------------------
Benchmark                         Time           CPU Iterations
----------------------------------------------------------------
TypeEqualsSimple                 11 ns         11 ns   65531752   178.723M items/s
TypeEqualsComplex                20 ns         20 ns   33939830   95.1497M items/s
TypeEqualsWithMetadata           31 ns         31 ns   22979555   62.4052M items/s
SchemaEquals                     40 ns         40 ns   17786532   48.1683M items/s
SchemaEqualsWithMetadata         46 ns         46 ns   15173158   41.3242M items/s
```