@pitrou 2019-08-01
@emkornfield @fsaintjacques @wesm Looking for design feedback on this.

Faster type equality may allow us to enable more runtime checks and make Arrow more user-friendly (see e.g. [ARROW-6038](https://issues.apache.org/jira/browse/ARROW-6038)).