@pitrou 2019-08-01
(note a similar optimization can be added to `Schema::Equals`)