@kszucs 2019-08-02
@fsaintjacques try running ursabot benchmark?