@pitrou 2019-08-05
> Do we want to keep type.h as small as possible, maybe the ComputeFingerprint could be a seperate VisitorType class?

I have no strong reference, but why not.