@pitrou 2019-08-05
Ok, I've updated the PR to handle `check_metadata` correctly and also handle schema equality. Will update benchmarks in the PR description.