@pitrou 2019-08-06
Note: still need to handle UnionType and ExtensionType, and I think that's all.