@pitrou 2019-08-06
The main downside here is that this adds a bit of complexity and fragility (`ComputeFingerprint` must be carefully implemented to match `Equals`). The upside of course is much faster type and (especially) schema comparison. Comparing schemas for hundreds of columns should actually have reasonable performance (under the hood it's probably a `memcmp` call).